pyproject.toml
[project]
name = "rag-for-small-company"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastembed>=0.7.4",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-fastembed>=0.5.0",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
]
